Ajustes na impressão boletim

+++ src/SME.SGP.Aplicacao/CasosDeUso/Relatorios/Boletim/BoletimEscolaAquiUseCase.cs
@@ -54,7 +54,7 @@ namespace SME.SGP.Aplicacao
                 var usuarioLogado = repositorioUsuario.ObterPorId(usuarioLogadoId); 
                 bool retorno;
 
-                if (relatorioBoletimEscolaAquiDto.Modelo == ModeloBoletim.Detalhado)
+                if (relatorioBoletimEscolaAquiDto.Modelo == (int)ModeloBoletim.Detalhado)
                     retorno = await mediator.Send(new GerarRelatorioCommand(TipoRelatorio.BoletimDetalhadoApp, relatorioBoletimEscolaAquiDto, usuarioLogado, RotasRabbitSgpRelatorios.RotaRelatoriosSolicitadosBoletimDetalhadoEscolaAqui,notificarErroUsuario:true));
                 else
                     retorno = await mediator.Send(new GerarRelatorioCommand(TipoRelatorio.Boletim, relatorioBoletimEscolaAquiDto, usuarioLogado, RotasRabbitSgpRelatorios.RotaRelatoriosSolicitadosBoletimEscolaAqui));
